alert

+++ clients/src/navbar/HeaderTeacher.js
@@ -1,6 +1,7 @@
 import React from "react";
 import "./HeaderTeacher.css";
 import LogoKU from '../assets/Logo.png';
+import Swal from 'sweetalert2'; // Import SweetAlert
 import Logout from '../assets/logout2.png';
 import { useNavigate } from "react-router-dom";
 
@@ -8,14 +9,35 @@ const HeaderTeacher = () => {
   const nav = useNavigate();
 
   const logout = async () => {
-    try {
-      nav("/");
-      sessionStorage.clear();
-      window.location.reload()
-    }
-    catch (error) {
-      console.error(error);
-    }
+    // Show confirmation popup
+    Swal.fire({
+      title: "ออกจากระบบ",
+      text: "คุณต้องการที่จะออกจากระบบใช่หรือไม่?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3CB371",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "ยืนยัน",
+      cancelButtonText: "ยกเลิก",
+      customClass: {
+        title: 'kanit-font',
+        content: 'kanit-font',
+        confirmButton: 'kanit-font',
+        cancelButton: 'kanit-font',
+        popup: 'kanit-font'
+      }
+    }).then((result) => {
+      if (result.isConfirmed) {
+        try {
+          nav("/");
+          sessionStorage.clear();
+          window.location.reload()
+        }
+        catch (error) {
+          console.error(error);
+        }
+      }
+    });
   };
 
   return (
